Feat/replace lunr to tinysearch (#119)
* feat: add tinysearch submodule

Added tinysearch as a submodule to the project. This will allow us to
integrate and manage the tinysearch repository within our main project
repository. The submodule is located at the path 'tinysearch' and points
to the repository at 'git@github.com:rikuson/tinysearch.git'.

* feat: replace lunr to tinysearch

Refactored the feed initialization to use a WebAssembly-based search
engine instead of parsing XML and using lunr. Updated the search
controller and auto-complete logic to work with the new feed structure.
This change improves performance and simplifies the codebase.

* feat: update CD

* fix: update unit test

* feat(ci): add caching for cargo dependencies

Added a step to cache cargo dependencies in the GitHub Actions workflow.
This will help speed up the build process by reusing previously downloaded
dependencies. The cache is keyed using the Cargo.lock file.